Moved from valueAsFloat to valueAsDouble and added accessor to Vector

+++ include/Analogs.h
@@ -39,8 +39,9 @@ public:
     ///
     /// Write all the subframes to a file by calling sequentially the write method of each subframe
     ///
-    void write(std::fstream &f,
-            std::vector<float> scaleFactors) const;
+    void write(
+            std::fstream &f,
+            std::vector<double> scaleFactors) const;
 
 
     //---- SUBFRAME ----//
