+++ include/Frame.h
@@ -41,9 +41,10 @@ public:
     ///
     /// Write the frame to a file by calling sequentially the write method for points and analogs
     ///
-    void write(std::fstream &f,
+    void write(
+            std::fstream &f,
             float pointScaleFactor,
-            std::vector<float> analogScaleFactors) const;
+            std::vector<double> analogScaleFactors) const;
 
 
     //---- POINTS ----//
