+++ include/Subframe.h
@@ -39,8 +39,9 @@ public:
     ///
     /// Write the subframe to a file by calling sequentially the write method of all of the analog channels
     ///
-    void write(std::fstream &f,
-            std::vector<float> scaleFactors) const;
+    void write(
+            std::fstream &f,
+            std::vector<double> scaleFactors) const;
 
 
     //---- CHANNELS ----//
