+++ src/Analogs.cpp
@@ -22,7 +22,7 @@ void ezc3d::DataNS::AnalogsNS::Analogs::print() const {
 
 void ezc3d::DataNS::AnalogsNS::Analogs::write(
         std::fstream &f,
-        std::vector<float> scaleFactors) const {
+        std::vector<double> scaleFactors) const {
     for (size_t i = 0; i < nbSubframes(); ++i) {
         subframe(i).write(f, scaleFactors);
     }
